backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
anthropic==0.27.0
httpx==0.27.0
redis[asyncio]==5.0.4
aiofiles==23.2.1
python-dotenv==1.0.1
pgvector==0.2.5
stripe==9.9.0
numpy==1.26.4
boto3==1.34.0
Pillow==10.3.0
slowapi==0.1.9
email-validator==2.1.1
apscheduler>=3.10.4
bcrypt==4.1.3
